Rtl test rewrites. Omstrukturering og mindre justeringer på koden i vurderingArbeidsmonster.js
