--- src/sider/eu_eøs/stegKomponenter/vurderingArbeidsmonster/landlinje/index.ts
@@ -1,3 +1,0 @@
-import LandLinje from "./landlinje";
-
-export default LandLinje;
