+++ src/sider/eu_eøs/stegKomponenter/vurderingArbeidsmonster/vurderingArbeidsmonster.test.js
@@ -1,22 +1,31 @@
 import React from "react";
+import { VurderingArbeidsmonster } from "./vurderingArbeidsmonster";
+import { renderWithProviders } from "../../../../ducks/test-utils/renderWithProviders";
 
-import * as KV from "../../../../kodeverk";
+jest.mock("nav-frontend-js-utils", () => ({
+  ...jest.requireActual("nav-frontend-js-utils"),
+  guid: () => "123",
+}));
 
-import { BOOLSK_STRING } from "../../../../constants";
-import MKV from "../../../../melosyskodeverk";
-import { lagAvklartfakta } from "../../../../felleskomponenter/stegvelger";
-
-import {
-  VurderingArbeidsmonster,
-  MarginaltArbeid,
-  CheckableLandLinje,
-  ConnectedCheckableLandLinje,
-  UkjenteEllerFlereEosLandLinje,
-} from "./vurderingArbeidsmonster";
-import LandLinje from "./landlinje";
-
-describe("VurderingVurderarbeidsland", () => {
+describe("VurderingArbeidsmonster", () => {
   let props = null;
+
+  const initialReduxState = {
+    form: {
+      soknad: {
+        initial: {
+          soknadsland: {
+            erUkjenteEllerAlleEosLand: true,
+          },
+        },
+        values: {
+          soknadsland: {
+            erUkjenteEllerAlleEosLand: true,
+          },
+        },
+      },
+    },
+  };
 
   beforeEach(() => {
     props = {
@@ -44,93 +53,10 @@ describe("VurderingVurderarbeidsland", () => {
     };
   });
 
-  it("viser Arbeidsmønstersteg uten å krasje", () => {
-    shallow(<VurderingArbeidsmonster {...props} />);
-  });
-});
-
-describe("MarginaltArbeid", () => {
-  let props = null;
-
-  beforeEach(() => {
-    props = {
-      arbeidsland: [
-        {
-          land: { kode: "DK", term: "Danmark" },
-          erLonnetArbeid: true,
-          erSelvstendigNaeringsvirksomhet: true,
-        },
-      ],
-      marginaltArbeid: [],
-      redigerbart: true,
-      oppdaterData: jest.fn(),
-      soknadsland: {
-        landkoder: ["DK"],
-        erUkjenteEllerAlleEosLand: false,
-      },
-    };
-  });
-
-  it("viser ConnectedCheckableLandLinje`r når erUkjenteEllerAlleEosLand er false", () => {
-    const marginaltArbeid = shallow(<MarginaltArbeid {...props} />);
-
-    const landLinjer = marginaltArbeid.find(ConnectedCheckableLandLinje);
-
-    expect(landLinjer).toHaveLength(1);
-  });
-
-  it("viser 1 UkjenteEllerFlereEosLandLinje når erUkjenteEllerAlleEosLand er true", () => {
-    props.soknadsland.erUkjenteEllerAlleEosLand = true;
-    const marginaltArbeid = shallow(<MarginaltArbeid {...props} />);
-
-    const landLinjer = marginaltArbeid.find(UkjenteEllerFlereEosLandLinje);
-
-    expect(landLinjer).toHaveLength(1);
-  });
-});
-
-describe("CheckableLandLinje", () => {
-  describe("ved klikk på checkbox", () => {
-    let props = null;
-
-    beforeEach(() => {
-      props = {
-        landKode: MKV.KTObjects.landkoder.find(({ kode }) => kode === MKV.Koder.landkoder.DE),
-        avklartMarginaltArbeidILand: { fakta: ["TRUE"] },
-        oppdaterData: jest.fn(),
-        redigerbart: true,
-        resetForm: jest.fn(),
-      };
-
-      const checkableLandLinje = shallow(<CheckableLandLinje {...props} />);
-      const landLinje = checkableLandLinje.find(LandLinje);
-
-      const checkboxOnCheck = landLinje.props().checkbox.onCheck;
-
-      checkboxOnCheck();
+  it("snapshot test", () => {
+    const { container } = renderWithProviders(<VurderingArbeidsmonster {...props} />, {
+      preloadedState: initialReduxState,
     });
-
-    it("lagrer marginalt arbeid avklartfakta", () => {
-      expect(props.oppdaterData).toHaveBeenCalledTimes(1);
-      expect(props.oppdaterData).toHaveBeenLastCalledWith(
-        lagAvklartfakta(MKV.Koder.avklartefaktatyper.MARGINALT_ARBEID, MKV.Koder.landkoder.DE, BOOLSK_STRING.USANN)
-      );
-    });
-
-    it("kaller resetForm for vedtak- og utpekformene", () => {
-      expect(props.resetForm).toHaveBeenCalledTimes(2);
-      expect(props.resetForm).toHaveBeenCalledWith(KV.Form.ARTIKKEL_13_X_VEDTAK);
-      expect(props.resetForm).toHaveBeenCalledWith(KV.Form.ARTIKKEL_13_UTPEKLAND);
-    });
-  });
-});
-
-describe("UkjenteEllerFlereEosLandLinje", () => {
-  it("viser en LandLinje med ikke-redigerbar checkbox", () => {
-    const ukjenteEllerFlereEosLandLinje = shallow(<UkjenteEllerFlereEosLandLinje />);
-    const landLinje = ukjenteEllerFlereEosLandLinje.find(LandLinje);
-
-    expect(landLinje).toHaveLength(1);
-    expect(landLinje.props().checkbox.redigerbart).toBe(false);
+    expect(container).toMatchSnapshot();
   });
 });
